trustgraph-vertexai/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trustgraph-vertexai"
dynamic = ["version"]
authors = [{name = "trustgraph.ai", email = "security@trustgraph.ai"}]
description = "TrustGraph provides a means to run a pipeline of flexible AI processing components in a flexible means to achieve a processing pipeline."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "trustgraph-base>=2.0,<2.1",
    "pulsar-client",
    "google-genai",
    "google-api-core",
    "prometheus-client",
    "anthropic",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/trustgraph-ai/trustgraph"

[project.scripts]
text-completion-googleaistudio = "trustgraph.model.text_completion.googleaistudio:run"
text-completion-vertexai = "trustgraph.model.text_completion.vertexai:run"

[tool.setuptools.packages.find]
include = ["trustgraph*"]

[tool.setuptools.dynamic]
version = {attr = "trustgraph.vertexai_version.__version__"}